+++ src/common/colors.js
@@ -3,5 +3,7 @@ export default {
   gray: '#767C84',
   lightBlue: '#F1F8FF',
   lightGray: '#F6F8FA',
+  red: '#FADCE0',
+  darkRed: '#881F26',
   white: '#FFFFFF',
 };
